REF: Refactor dataset reader to open multiple files
